Module SHA256.HMAC

val digest : key:Stdlib.Bytes.t -> msg:Stdlib.Bytes.t -> Stdlib.Bytes.t

digest key msg HMAC based on the hash algorithm of the module. The size of the produced tag is the same as the size of the hash digest.